renamed ui/notifications to userprofile/UserProfileViewModel and NotificationsViewModel to UserProfileViewModel
