+++ app/src/main/java/com/example/instafameproj/ui/userprofile/UserProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.example.instafameproj.ui.notifications
+package com.example.instafameproj.ui.userprofile
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -23,7 +23,7 @@ class UserProfileFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
         val notificationsViewModel =
-            ViewModelProvider(this).get(NotificationsViewModel::class.java)
+            ViewModelProvider(this).get(UserProfileViewModel::class.java)
 
         _binding = FragmentNotificationsBinding.inflate(inflater, container, false)
         val root: View = binding.root
