+++ app/src/main/java/com/example/instafameproj/ui/userprofile/UserProfileViewModel.kt
@@ -1,10 +1,10 @@
-package com.example.instafameproj.ui.notifications
+package com.example.instafameproj.ui.userprofile
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class NotificationsViewModel : ViewModel() {
+class UserProfileViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is User Profile Fragment"
